Patient-Service/src/main/java/com/arivanamin/healthcare/backend/patient/application/request/CreatePatientRequest.java
package com.arivanamin.healthcare.backend.patient.application.request;

import com.arivanamin.healthcare.backend.core.domain.gender.Gender;
import com.arivanamin.healthcare.backend.patient.core.entity.Patient;
import lombok.*;
import org.modelmapper.ModelMapper;

import java.time.LocalDate;
import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CreatePatientRequest {
    
    UUID id;
    String firstName;
    String lastName;
    String email;
    LocalDate dateOfBirth;
    Gender gender;
    String address;
    
    ModelMapper mapper = new ModelMapper();
    
    public Patient toEntity () {
        return mapper.map(this, Patient.class);
    }
}
